Feature/#20&38 (#40)
